refactor theme context and save theme to firebase
