+++ src/widget/Loading/Loading.tsx
@@ -1,10 +1,10 @@
 import React, { useContext } from 'react';
 import './Loading.scss';
-import { CurrentThemeContext } from '../../App';
 import cn from 'classnames';
+import { CurrentThemeContext } from "../Widget/Widget";
 
 const Loading: React.FC = () => {
-    const isLightTheme = useContext(CurrentThemeContext);
+    const { isLightTheme } = useContext(CurrentThemeContext);
 
     return (
         <div className={cn('lds-default', isLightTheme ? 'light-lds' : 'dark-lds')}>
