+++ src/widget/Widget/Widget.tsx
@@ -1,7 +1,6 @@
 import React, { useContext, useEffect, useState } from 'react';
 import ChangeGroupButton from '../ChangeGroupButton/ChangeGroupButton';
 import './Widget.scss';
-import { CurrentThemeContext } from '../../App';
 import FavoriteIcon from '../../assets/icons/FavoriteIcon';
 import PeopleIcon from '../../assets/icons/PeopleIcon';
 import SettingsIcon from '../../assets/icons/SettingsIcon';
@@ -30,11 +29,20 @@ const FAVOURITES_COUNT = EMOJIS_IN_ROW * 7;
 enum StatesKeys {
     CurrentGroupName = 'currentGroupName',
     RecentEmojis = 'recentEmojis',
-    FavouritesEmojis = 'favouritesEmojis'
+    FavouritesEmojis = 'favouritesEmojis',
+    IsLightTheme = 'isLightTheme'
 }
 
+type ThemeContextType = { isLightTheme: boolean; dispatchChangeTheme: (value: boolean) => void };
+
+export const CurrentThemeContext = React.createContext<ThemeContextType>({
+    [StatesKeys.IsLightTheme]: true,
+    dispatchChangeTheme: () => {}
+});
+
 const Widget: React.FC = () => {
-    const isLightTheme = useContext(CurrentThemeContext);
+    const [isLightTheme, setIsLightTheme] = useState<boolean>(loadWidgetState(StatesKeys.IsLightTheme, true));
+
     const { firestore } = useContext(FirebaseContext);
     const { auth } = useContext(FirebaseContext);
 
@@ -54,13 +62,17 @@ const Widget: React.FC = () => {
     useEffect(() => {
         if (user) {
             getDoc(doc(firestore, 'users', user?.uid)).then(data => {
+                const light = data.get(StatesKeys.IsLightTheme);
                 const favourites = data.get(StatesKeys.FavouritesEmojis);
                 const recent = data.get(StatesKeys.RecentEmojis);
-                if (favourites && recent) {
+                if (light !== null && favourites && recent) {
+                    setIsLightTheme(light);
                     setFavouritesEmojis(favourites);
                     setRecentEmojis(recent);
                 } else {
+                    // TODO вынести в отдельную функцию и пофиксить баг с переключателем
                     setDoc(doc(firestore, 'users', user.uid), {
+                        [StatesKeys.IsLightTheme]: isLightTheme,
                         [StatesKeys.RecentEmojis]: recentEmojis,
                         [StatesKeys.FavouritesEmojis]: favouritesEmojis
                     }).then(_ => console.log('Save recent and favourites to firebase after creating account'));
@@ -73,20 +85,22 @@ const Widget: React.FC = () => {
         localStorage.setItem(StatesKeys.CurrentGroupName, JSON.stringify(currentGroupName));
         localStorage.setItem(StatesKeys.RecentEmojis, JSON.stringify(recentEmojis));
         localStorage.setItem(StatesKeys.FavouritesEmojis, JSON.stringify(favouritesEmojis));
-    }, [currentGroupName, recentEmojis, favouritesEmojis]);
+        localStorage.setItem(StatesKeys.IsLightTheme, JSON.stringify(isLightTheme));
+    }, [currentGroupName, recentEmojis, favouritesEmojis, isLightTheme]);
 
     useEffect(() => {
         if (user) {
             setDoc(doc(firestore, 'users', user.uid), {
+                [StatesKeys.IsLightTheme]: isLightTheme,
                 [StatesKeys.RecentEmojis]: recentEmojis,
                 [StatesKeys.FavouritesEmojis]: favouritesEmojis
             }).then(_ => console.log('Save recent and favourites to firebase'));
         }
-    }, [recentEmojis, favouritesEmojis]); // eslint-disable-line react-hooks/exhaustive-deps
+    }, [recentEmojis, favouritesEmojis, isLightTheme]); // eslint-disable-line react-hooks/exhaustive-deps
 
     function loadWidgetState(key: string, defaultValue: any) {
         const state = JSON.parse(localStorage.getItem(key) as string);
-        if (state) return state;
+        if (state !== null && typeof state !== 'undefined') return state;
         return defaultValue;
     }
 
@@ -141,44 +155,51 @@ const Widget: React.FC = () => {
     ];
 
     return (
-        <div className={cn('widget', isLightTheme ? 'light-widget' : 'dark-widget')}>
-            <div className={'buttons-wrapper'}>
-                {icons.map((icon, index) => (
-                    <ChangeGroupButton
-                        key={`ChangeGroupButton${index}`}
-                        groupName={emojiGroups[index]}
-                        icon={icon}
-                        onClick={changeCurrentGroupData(emojiGroups[index])}
-                        isActive={emojiGroups[index] === currentGroupName}
+        <CurrentThemeContext.Provider
+            value={{
+                [StatesKeys.IsLightTheme]: isLightTheme,
+                dispatchChangeTheme: (value: boolean) => setIsLightTheme(value)
+            }}
+        >
+            <div className={cn('widget', isLightTheme ? 'light-widget' : 'dark-widget')}>
+                <div className={'buttons-wrapper'}>
+                    {icons.map((icon, index) => (
+                        <ChangeGroupButton
+                            key={`ChangeGroupButton${index}`}
+                            groupName={emojiGroups[index]}
+                            icon={icon}
+                            onClick={changeCurrentGroupData(emojiGroups[index])}
+                            isActive={emojiGroups[index] === currentGroupName}
+                        />
+                    ))}
+                </div>
+                {currentGroupName === Groups.Favourites ? (
+                    <FavouritesGroup
+                        recentEmojis={isSearching ? searchedEmojis : recentEmojis}
+                        favouritesEmojis={favouritesEmojis}
+                        updateSearchedGroup={updateSearchedGroup}
+                        updateRecentEmojis={updateRecentEmojis}
+                        isSearching={isSearching}
+                        setIsSearching={setIsSearching}
                     />
-                ))}
+                ) : currentGroupName === Groups.Settings ? (
+                    <SettingsGroup />
+                ) : (
+                    <BasicGroup
+                        groupName={isSearching ? Groups.Searched : currentGroupName}
+                        groupEmojis={isSearching ? searchedEmojis : currentGroupEmojis}
+                        updateSearchedGroup={updateSearchedGroup}
+                        updateRecentEmojis={updateRecentEmojis}
+                        setIsSearching={setIsSearching}
+                    />
+                )}
+                <ContextMenu
+                    addFavourite={addFavouriteEmoji}
+                    removeFavourite={removeFavouriteEmoji}
+                    updateRecent={updateRecentEmojis}
+                />
             </div>
-            {currentGroupName === Groups.Favourites ? (
-                <FavouritesGroup
-                    recentEmojis={isSearching ? searchedEmojis : recentEmojis}
-                    favouritesEmojis={favouritesEmojis}
-                    updateSearchedGroup={updateSearchedGroup}
-                    updateRecentEmojis={updateRecentEmojis}
-                    isSearching={isSearching}
-                    setIsSearching={setIsSearching}
-                />
-            ) : currentGroupName === Groups.Settings ? (
-                <SettingsGroup />
-            ) : (
-                <BasicGroup
-                    groupName={isSearching ? Groups.Searched : currentGroupName}
-                    groupEmojis={isSearching ? searchedEmojis : currentGroupEmojis}
-                    updateSearchedGroup={updateSearchedGroup}
-                    updateRecentEmojis={updateRecentEmojis}
-                    setIsSearching={setIsSearching}
-                />
-            )}
-            <ContextMenu
-                addFavourite={addFavouriteEmoji}
-                removeFavourite={removeFavouriteEmoji}
-                updateRecent={updateRecentEmojis}
-            />
-        </div>
+        </CurrentThemeContext.Provider>
     );
 };
